Fixes varios (#60)
* rama nueva

* resuelve un conflicto en el buffer

* se reactiva la musica meme y la pantalla de desconexion

* fix de muere uno y seg fault en server

* modifica algunas variables y metodos realcionados con el atributo connectionOn en Client y disconnected en GameClient que parecian o mal manejado por falta de mutex o mal manejados por la logica de la desconexion del juego

* solo quita lo  no bloqueante y agrega un metodo para que ni bien hay error de sendo o recive del lado server que se corte la conexion con ese cliente

* habia faltado quitar un metodo mas de no bloqueante. Se agrega para que en el listen se agregue una UserConnection SOLO si el socket es un int positivo (es decir, no hubo erro)

* agrega el KEEP_ALIVE flag

* quita el KEEP_ALIVE y vuelve a poner el no bloqueante pero del lado cliente

* agregar el non blocking para server tamb

* corrige y mejora el parseo de mensajes entrantes para que no reciva basura

* agrega el checkeo por EAGAIN en caso de no haber datos para send o recv (POR ESTO FALLABA CONSTANTEMENTE EL READ)

* homogeiniza las configs sp, 2p, 4p  cambiando solo los amounts

* corrige la logica de manejo de errores en el recv de server y client y ademas quita la linea que hacia setConnectionOff en el metodo error de Client para evitar desconexiones inesperadas, reduciendo el metodo error a un simple print por log

* mejora el detalle de errores de sending en Client y Server

* retoca un logeo de Error en sending y reading en Server y Client. Corrige un logeo de paths en el XMLParser y homegeiniza los paths en la config default para evitar problemas de mapeo en el caso default que deberia ser el mas estable

* mejora el logeo de las conexiones y reconexiones y de la reproduccion de la musica de fondo

* retoque minimo en logeo de desconexiones, reconexiones y muertes

* agrega logeo de seniales recibidas desde server a cliente

* agrega para que Soundable y Renderable no guarde en el map un TextureWrapper o un SoundWrapper asociado a un path vacio

* modifica un poco mas la logica de manejo de errores en el send y recv de Server y Client

* CAMBIOS MINIMOS: hace mas legible el game loop en game client

* corrige bugs de modo test de waiting screen y que un muerto pueda setearlo

* agrega logeo de mensajes corruptos al parser y fix de un bug en el waitingForTeammates

* mejora el parser (faltaba un contemplar un caso) y ademas logea y printea los mensajes corruptos

* agrega mejores prints de los mensajes corruptos en Client

* se corrige un poco mas la logica del send y recv en el manejo de errores para que en todos los casos haga un return adecaudo y se modifican algunos parametros de send y recv

* fix importante del ultimo commit. Estaba usando el socketFD equivocado

* fix importante del ultimo commit. Estaba usando el socketFD equivocado (2)

* prints

* mas prints

* pone en bloqueante otra vez el send y recv en server y cliente

* cambia la condicion de corte en el while de send y recv en server y client

* parece que las conexiones andan de 10 y sin mensajes corruptos. Este es el commit antes de correr todas las pruebas de conexion

* se vuelve a hacer no blockeante. Antes habian pasado todas la pruebas salvo las relacionadas con desconexion de cable

* Ahora si hace no bloqueante (me habia faltado una linea) y fix de boton de musica

* fix para que compile (ups)

* lo hace bloqueante denuevo

* prueba con non block pero con mas cosas

* prueba un cambio en parametros de non block

* mini fix en send y recv

* add cantidad de vidas restantes

* desconexiones andando bien salvo el detalle de la tardanza al desconectarse

* agrega un checkeador en controller

* probando cosas para el input

* ESTE ANDUVO BIEN CON FRAN

* parece que arregla que se trabe movimiento por el input que no le llegaba y arregla el bug de que quedavisualmente arriba al saltar

* agrega mutex en lista players en entityManager

* agrega mas usos de muteex en entityManager por que no se habia contemplado todos los usos de la lista players

* AHORA SI se arregla el salto que queda como volando y tamb se arregla el ruido con delay y lo de que queda trabado por un seg desp de haber saltado

* se arregla el bug de que suena la patada desp de haber pegado cuando le pegas al aire

* se resuelve un problema visual y de sonido de que al tercer golpe con arma cuando el arma se rompia parecia que el tercer golpe habia sido dado con punio

* probando otra implementacion del while del recv del Cliente

* probando agregarle un sleep de 2 segs antes de la respuesta del server ante el login

* puse prints

* mismo fix que el client pero lado server

* lo ultimo que hicimos

* fix? del esperando respuesta del lado server?

* ahora si que si se fixea el salto que queda trabado en el aire o en el piso

* mejora el parseo de mensajes en el recv porque antes no se estaba limpiando caracteres invalidos por izquierda del mensaje correcto

* fix de la hitting box que se rompia en el lvl 2

* heehee agregado

* heehee fix

* se emparcha el bug del salto que la animacion del salto quedaba volando

* sin leaks de memoria

* agrega para que si se tiene un arma en la mano que no se pueda agarrar otra

* use numbers from config

* fix de seg fault y restaura enemigos danio y spawn lvl 1 y idem el final boss

+++ src/game/GameServer.cpp
@@ -32,40 +32,40 @@ void GameServer::start() {
 void GameServer::gameLoop(){
 
     sendGameStartedMessage();
-    while (isOn() && levelBuilder->hasNextLevel() && thereIsAtLeastOnePlayerAliveAndConnected()) {
+    while (isOn() && levelBuilder->hasNextLevel() && thereIsAtLeastOnePlayerAliveAndConnected()) { //levels loop
 
         levelBuilder->loadNext();
         LogManager::logInfo("=======================================");
         LogManager::logInfo("[GAME]: se inicia game loop de este nivel");
 
-        while (isOn() && !levelBuilder->levelFinished() && thereIsAtLeastOnePlayerAliveAndConnected()) {
+        while (isOn() && !levelBuilder->levelFinished() && thereIsAtLeastOnePlayerAliveAndConnected()) { //game loop
             update();
             sendUpdate();
             usleep(SLEEP_TIME);
         }
-         if (thereIsAtLeastOnePlayerAliveAndConnected()) {
+        if (thereIsAtLeastOnePlayerAliveAndConnected()) { //score screen
             sceneDirector->initScoreScreen(entityManager->getPlayers(),loggedPlayersUserByID);
             sceneDirector->sendScoreScreen(server);
              usleep(WAIT_TIME);
         }
-
         LogManager::logInfo("[GAME]: Nivel terminado");
         LogManager::logInfo("=======================================");
     }
-    if (notAllPlayersDisconnected() ) {
+    if (notAllPlayersDisconnected() ) { //end screen
         sendEndMessage();
         sceneDirector->initEndOfGameScreen();
-        sceneDirector->sendEndOfGameScreen(server);
-        usleep(WAIT_TIME);
-    } 
-    on = false;
+
+        for (int i = 0; i < END_SCREEN_SEND_AMOUNT; i++){ //esto es porque la ser de lo ultimo que manda el server hay que mandarlo varias veces para que llegue
+            sceneDirector->sendEndOfGameScreen(server);
+        }
+    }
+    end();
     server->stopListening();
 }
 
 void GameServer::update() {
     entityManager->update();
     levelBuilder->update();
-    controller->clearAllInputs();
 }
 
 void GameServer::sendUpdate() {
@@ -87,6 +87,7 @@ void GameServer::sendUpdate() {
 void GameServer::handleLogin(const std::string& user, const std::string& pass, int userId){
 
     string toSendMessage;
+    //usleep(2000000);
 
     if (credentialsAreValid(user,pass)){
         if (userAlreadyLoggedIn(user)){
@@ -124,7 +125,7 @@ void GameServer::reemplazePreviousIDWith(int oldID, int newID) {
 }
 
 void GameServer::reciveNewInput(tuple<Action,int> input){
-    return controller->setInput(input);
+    return entityManager->setInput(input);
 }
 
 bool GameServer::isActive(){
@@ -141,16 +142,20 @@ bool GameServer::thereIsAtLeastOnePlayerAliveAndConnected() {
 
 void GameServer::recibeTestModeSignal() {
 
-    inTestMode = !inTestMode;
+    if (entityManager != nullptr){
 
-    if (inTestMode){
-        entityManager->setTestMode();
-        LogManager::logInfo("[GAME]: modo test activado");
+        inTestMode = !inTestMode;
+
+        if (inTestMode){
+            entityManager->setTestMode();
+            LogManager::logInfo("[GAME]: modo test activado");
+        }
+        else{
+            entityManager->removeTestMode();
+            LogManager::logInfo("[GAME]: modo test desactivado");
+        }
     }
-    else{
-        entityManager->removeTestMode();
-        LogManager::logInfo("[GAME]: modo test desactivado");
-    }
+
 }
 
 //SERVER RELATED
@@ -185,10 +190,10 @@ void GameServer::connectionLostWith(int id){
 
     bool logged = loggedPlayersUserByID.count(id);
     bool dead = deadPlayers.count(id);
+    string name = "UNKNOWN";
 
     if (logged){
-
-        string name = loggedPlayersUserByID.at(id).name;
+        name = loggedPlayersUserByID.at(id).name;
         disconectedPlayers.insert({name,id});
 
         if (!dead){
@@ -198,6 +203,7 @@ void GameServer::connectionLostWith(int id){
     if (entityManager != nullptr){
         entityManager->disconectPlayerByID(id);
     }
+    LogManager::logInfo("[GAME]: se proceso al jugador desconectado: " + to_string(id) + " | " + name);
 }
 
 //CONTROLLER RELATED
@@ -211,6 +217,7 @@ void GameServer::notifyPlayerDied(int id) {
     conectedAndPlayingPlayersAmount --;
     controller->sendPlayerDiedMessage(server,id);
     deadPlayers.insert({id,loggedPlayersUserByID.at(id).name});
+    LogManager::logInfo("[GAME]: se proceso al jugador muerto: " + to_string(id) + " | " + loggedPlayersUserByID.at(id).name);
 }
 
 void GameServer::sendGameStartedMessage() {
@@ -261,8 +268,9 @@ string GameServer::processConectionAndEmitSuccesMessage(const string& name, cons
     loggedPlayersUserByID.insert({ id, user });
     loggedPlayersIDbyUser.insert({user.name,id});
     addNewIDToGame(id);
-    server->client_noBlock(id);
+    //server->client_noBlock(id);
 
+    LogManager::logInfo("[GAME]: se proceso al jugador conectado: " + to_string(id) + " | " + user.name);
     return controller->getSuccesfullLoginMessage(user.color, id);
 }
 
@@ -293,9 +301,10 @@ void GameServer::processReconectionAndEmitSuccesMessage(const string& name, int 
         conectedAndPlayingPlayersAmount++;
     }
 
-    server->client_noBlock(newID);
+    //server->client_noBlock(newID);
     server->setToSendToSpecific(controller->getSuccesfullLoginMessage(user.color,newID),newID);
     server->setToSendToSpecific(controller->getGameStartedMessage(),newID);
+    LogManager::logInfo("[GAME]: se proceso al jugador reconectado: " + to_string(newID) + " | " + user.name);
 }
 
 string GameServer::getNewColor() {
